Nsec/preparations (#202)
This PR has many code refactorings that introduce no behavioural change, but are very important for many reasons, first of all because ++ is a performance anti-pattern in an immutable language and this construct is heavily abused in this repo, so a lot of the code is refactored to behave equally with better practices. See https://www.erlang.org/doc/system/listhandling.html for details.

Other changes include removing unnecessary fully-qualified function calls, fixing type declarations for types that didn't exist, using lists:flatmap/2 instead of lists:flatten(lists:map/2), reducing the nesting of many function calls to aid readability, etc.

* Update dns_erlang to have NXNAME

* Refactor erldns_edns:get_opts/1 list handling

* Refactor erldns_resolver list handling and nesting

* Fix types in erlnds_records

* Refactor erldns_handler types, state, docs, and gen_server

* Refactor erldns_dnssec types and list handling