Event handler (#207)
This process is a single global process where the system register handlers. These handlers react on all metrics and log events, and therefore a single process becomes responsible for handling with virtually all metrics and logs of a quite big system, meaning, a bottleneck and a SPOF. Also, having this process deal with logs also means logs lose information about their calls, like stacktraces or file and line positions.

Instead, at the points of log and metric generations, these can be incremented locally. Performance is not a problem, metrics usually just increment a counter in an ETS table, which is very likely faster than passing a message with a big payload to a bottlenecked process. Likewise logs, most logs will only be filtered out against a log-level, which in the case of OTP Logger is indeed faster than any message passing even.