Rework packet cache
Use `segmented_cache` for the logic.

This module is already heavily optimised for unbounded concurrency, and
for very cheap bookkeeping (keeping track of expired records) at the
expense of granularity. That is, instead of keeping the exact timestamp
of each and every record and then sweeping by exact values, it keeps
groups of records with a "younger than" timestamp, at sweeps the groups
entirely, at the expense of killing records earlier than expected, but
never later.