Create packet pipelines (#224)
Uses a purely functional piping idiom, see erldns_pipeline for details. Splits all the logic of erldns_handler:handle/2 into a bunch of logically encapsulated steps that can be applied one after another in a configurable and plugable way.

erldns_resolver remains mostly untouched, though it feels like it's a really big one and could be split into smaller pipes, I haven't touched it for now, as the intention was only to create a framework that allows extensibility. For example dnssec logic could be extracted into a new pipeline handler.