Network stack reimplementation. (#225)
It reimplements the TCP and UDP stack with kernel support for REUSEADDR/REUSEPORT and alarm systems for overload, (on my hardware) the TCP stack gets a 2x performance improvement, the UDP stack gets a ~15% performance improvement and a more balanced CPU consumption, for a much smaller codebase. Coverage is ~90% on all modules related to this change, and documentation should also be extensive and helpful.

It doesn't implement backpressure and load protection on the acceptor pool yet.