Correct Ranch parallelism
This flag is per configured connection supervisor and not global,
hence we don't need to multiply here by parallelism.

The value chosen is by the following calculation: if a light request
takes 1ms, and we queue N of them in a scheduler, the last one would be
delayed by N ms, which would probably be about to hit the ingress
timeout. So that's the maximum number of requests per scheduler allowed.