Rework erldns_handler to be performant and type-safe (#258)
Uses persistent_term instead of ETS tables, and callbacks are preallocated instead of resolved at runtime again and again. Same optimisations than zone codecs or pipelines had.

This will be useful when reworking specially the resolver, as currently so much code internal to the handlers is scattered there and in the dnssec module. It also clearly defines a behaviour for this module, as callbacks were not specified.