Fix bug with SOA records not updating correctly (#266)
Note that the SOA record is stored twice: once in the zones table (the
`authority` section of the `#zone{}` record), and another one as a
standalone RRSet in the typed records table. And then the resolution
algorithm would use the RRSet in the typed records table as the answer
for an explicit query to SOA records but the cached ones on the zone
record for any other query that required including SOA records. This
would be an issue for DNSSEC, because for example an NSEC answer would
include the RRSIG record of the _new_ SOA record, but with the _old_ SOA
record, hence triggering a validation failure.

Note that this code was available in v7, and it was lost just very unfortunately. The PR comes with a regression test.