Do not put empty RRSIG in the cache (#278)
When deleting records, an empty list could be a value for an RRSIG record that
would be put in the cache. This could happen even if the zone had no DNSSEC.
Because of how the "getters" from ETS worked (for example pattern_zone_dname/2)
by appending the resulting record list, this empty RRSIG record would not
appear usually. However, it would get counted as an existing record, which could
lead to bugs.